requirements.txt
openpyxl~=3.1.1
requests~=2.28.2
PyYAML~=6.0